Weapons from GS to KPCF
Added the weapons from the gearscripts into KPCF. 

Might tinker with this some more to see if the blacklist works so there arent loads of different ammo types. Not that I actually think that will present any issues, but if its possible its nice to be safe. Will only tinker with that after you push this PR through so as not to accident anything into main branch.